Se agregaron eps y rh al estudiante

+++ src/main/java/com/app/akdemy/interfacesServices/IEstudianteService.java
@@ -11,6 +11,6 @@ public interface IEstudianteService {
 
     public List<Estudiante> listarEstudiantes();
 
-    public Estudiante buscarPorId(int id);
+    public Estudiante buscarPorId(long id);
     
 }
